Documentation for response classes. Move some classes to specific folders
