+++ src/CustomExceptionMiddleware.csproj
@@ -30,9 +30,9 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<None Include="..\icon.png" Pack="true" PackagePath=""/>
-		<None Include="..\README-nuget.md" Pack="true" PackagePath="\"/>
-		<None Include="..\LICENSE" Pack="true" PackagePath=""/>
+		<None Include="..\icon.png" Pack="true" PackagePath="" />
+		<None Include="..\README-nuget.md" Pack="true" PackagePath="\" />
+		<None Include="..\LICENSE" Pack="true" PackagePath="" />
 	</ItemGroup>
 
 </Project>
